requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cloudinary==1.41.0
Django==5.1.3
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
Pillow==11.0.0
pytz==2021.1
sqlparse==0.5.2
